[New] Wave export can now also render by instrument (i.e. one wave file per instrument). [Imp] General tab: The "output to" dropdown list also shows the actual plugin name (instead of just the library name) now. [Mod] Updated release notes [Mod] Some datatype changes in r574 were wrong in theory (but did not change anything because of the way C++ work), but I reverted them anyway. [Ref] Some more refactoring
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@602 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/misc_util.h
@@ -112,7 +112,7 @@ template <size_t size>
 inline void SanitizeFilename(char (&buffer)[size])
 {
 	STATIC_ASSERT(size > 0);
-	for(int i = 0; i < size; i++)
+	for(size_t i = 0; i < size; i++)
 	{
 		if(	buffer[i] == '\\' ||
 			buffer[i] == '\"' ||
@@ -123,7 +123,7 @@ inline void SanitizeFilename(char (&buffer)[size])
 			buffer[i] == '>'  ||
 			buffer[i] == '*')
 		{
-			for(int j = i + 1; j < size; j++)
+			for(size_t j = i + 1; j < size; j++)
 			{
 				buffer[j - 1] = buffer[j];
 			}
